requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --python-version 3.10 -o requirements.txt
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.14.3
    # via pageindex (pyproject.toml)
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via pdfminer-six
cryptography==46.0.7
    # via pdfminer-six
et-xmlfile==2.0.0
    # via openpyxl
lxml==6.1.0
    # via
    #   pageindex (pyproject.toml)
    #   python-docx
numpy==2.2.6
    # via pandas
openpyxl==3.1.5
    # via pageindex (pyproject.toml)
packaging==26.1
    # via pytesseract
pandas==2.3.3
    # via pageindex (pyproject.toml)
pdf2image==1.17.0
    # via pageindex (pyproject.toml)
pdfminer-six==20251230
    # via pdfplumber
pdfplumber==0.11.9
    # via pageindex (pyproject.toml)
pillow==12.2.0
    # via
    #   pageindex (pyproject.toml)
    #   pdf2image
    #   pdfplumber
    #   pytesseract
pycparser==3.0
    # via cffi
pydantic==2.13.2
    # via pageindex (pyproject.toml)
pydantic-core==2.46.2
    # via pydantic
pypdfium2==5.7.0
    # via pdfplumber
pytesseract==0.3.13
    # via pageindex (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via pageindex (pyproject.toml)
pytz==2026.1.post1
    # via pandas
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   cryptography
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2026.1
    # via pandas
